opaque-origins/opaque-worker.html

<!DOCTYPE html>
<meta charset=utf-8>
<script src="../resources/testharness.js"></script>
<!--<script src="/resources/testharnessreport.js"></script>-->
<body>
<script>
async_test(function(t) {
    self.addEventListener('message', t.step_func(function(e) {
        if (e.source == window[0]) {
          assert_equals(e.data, 'from page');
          t.done();
        }
      }));
  }, 'A worker is same origin with its opaque origin creator');
async_test(function(t) {
    self.addEventListener('message', t.step_func(function(e) {
        if (e.source == window[1]) {
          assert_equals(e.data, 3);
          t.done();
        }
      }));
  }, 'Shared workers work in opaque origins');
</script>
<iframe id='fetch' sandbox="allow-scripts" src="resources/sandboxed.html"></iframe>
<iframe id='sharedworker' sandbox="allow-scripts" src="resources/sandboxed-shared-worker.html"></iframe>
</body>
